+++ programs/coinflip/src/instructions/complete_game.rs
@@ -18,7 +18,7 @@ pub fn handler(
     // ===============================
 
     require!(
-        game.waiting_for_oracle(oracle.oracle_buffer_time as u64, current_time),
+        game.waiting_for_oracle(oracle.oracle_buffer_time, current_time),
         ErrorCode::GameNotReadyForOracle
     );
